Add server check for support form uploads
The client side already has validation for trying to attach non-images,
so this is mostly an anti-spambot measure to prevent people from
throwing viruses into the ticketing system (even though this would
require a support person to intentionally download and open them).

The `filename` and `mime_type` of a Plug.Upload are both provided by the
user and thus untrusted, but since we ultimately attach the upload to an
email using its `filename`, that is what will control how the
recipient's computer tries to open it.