Fix Jest watch mode not working
This was a bug that was fixed in a newer version of Jest, so it only
required updating. This also adds a `npm run test:jest:watch` shortcut.

The Jest branch coverage threshold had to be adjusted down as part of
this, since Jest 24.1.0 fixed an issue where test setup files were
included in the code coverage stats. This commit does not change our
"real" code coverage.