Create JourneyList API response
Using minimal information, return a complete list of journeys for a given route provided the direction, date, and origin in an API request. Journeys are "cleaned" to avoid encoding errors when returned as JSON. These journeys may include trips that are "added in" for the current service/day (those with predictions but no schedules).