Fix alert badge not appearing on combined GL page
This occurred because the custom `green_alerts` function used here did
not populate the `all_alerts_count` assign, which is used on the other
line pages to display the Alerts tab badge.

Since `green_alerts` appears to be a subset of `assign_alerts` which
would normally assign the alert count, and that function seems to work
fine on the combined Green Line page, we can replace it wholesale.