Add maps to shuttles tab
Adds two maps to the shuttles tab to show shuttle service overview and a detail map zoomed in to a selected stop. Shows relevant subway line along with shuttle diversions. Shows relevant subway and shuttle stops with a minimal tooltip. Configures interaction between the direction and stop selectors and the corresponding lines and stops on the maps.

Additional changes:
* Modifies the existing leaflet Map react component to render dotted lines
* Removes (unused) `dotted?` attribute value from trip_plan map, thus allowing the Trip Planner results to continue to display a solid line for walking portions of a trip.
* Adds color key to the Route and populates with appropriate color
* Adds new icon to represent subway stops adversely affected by shuttle service