Fix shuttle data including normal route shapes
This occurred because, when we run normal service for part of a day and
shuttle service for the rest, all of those trips are grouped under one
Service with a `schedule_typicality` of 4, despite only part of the day
being "atypical". We should instead look at the RoutePattern, which has
a typicality of 1 for the normal un-shuttled trips.