Fix disabled scroll bug with Schedule Finder
This fixes an issue where clicking a link from the Schedule Finder modal
(e.g. one of the stops in the schedule) would result in the destination
page, and possibly later visited pages, being un-scrollable until a full
page reload was performed.

This was caused by the modal setting a `modal-open` class on the `html`
element, which then persisted during Turbolinks navigation, which only
replaces the `body`. It appears we intended to disable Turbolinks on the
pages that open the modal, but the logic was wrong and would keep it
enabled if it had ever been enabled on a previous page (without any full
page reloads in between).