Relocate and improve default service selection
Lets Elixir determine the service that best matches today, and passes the ID for this default service over to React. Tags serviecs as valid for today's date or not, then determines a best match based on several factors from this list of candidates. If today's date is in the :added_dates list, that is the best match, followed by any service that matches today's day number, with a final fallback of the first service in the valid-today candidates list, which is presorted to be most relevant.
- Removes all react-side logic for determining default service selection
- Call JourneyList.build/5 with current_time depending on it being today or not
- add test coverage that checks for default service assignment
- adds dependency injection for date and testing for better default service selection
- fixed after midnight bug by using `Util.service_date/1` on current date_time
- fixed failing after midnight tests by using fixed, future service date
- fixed other failing after midnight test by using `Util.service_date/1` instead of `Util.now/0`