Fix combo GL shuttle map not including all shapes
In the case where shuttle data was requested for multiple routes but
only some of those routes had a shuttle diversion, the returned stops
would be from all requested routes, but the shapes would only be those
corresponding to the diverted routes.

This occurred because we were building the shapes based on trips that
had an atypical route pattern, but for routes unaffected by a diversion,
none of the trips would have an atypical route pattern.

We resolve this by adding an extra set of trips to the shape building
process, consisting of those which 1) are typical, and 2) do not belong
to the routes affected by active shuttle alerts.