Rework and clean up Backstop tasks
The WIREMOCK_PROXY variable causes the backend to add a `X-WM-Proxy-Url`
header to external requests, which in turn enables the wildcard `proxy`
rule in Wiremock and causes un-matched requests to be forwarded to the
internet.

Previously we enabled this for routine Backstop test runs, but this
meant that if some requests hadn't been properly recorded, they would
silently use live data and cause spurious CI failures. This change adds
an explicit `:record` sub-task for Backstop that enables this variable,
and leaves it disabled otherwise, causing Wiremock to return an HTML 404
for un-recorded requests.

This also removes some `mix` tasks that were related to Backstop; they
seemingly duplicate functionality already available in Backstop itself.