Find holidays by typicality for schedule finder
We'd been using a heuristic to decide which services in the schedule
finder represented a holiday, to wit, we assumed any service applying to
just a single day was holiday service. For any number of reasons, this
did not produce satisfactory results.

Going forward, holidays are going to be represented in a more consistent
way in the API, allowing us to recognize them by their typicality, to
determine which holidays they apply to by their added_dates, and the
names of those holidays via added_dates_notes.