Hide animated icon from Backstop comparisons
Since this blinking icon would animate for an inconsistent amount of
time before Backstop took a screenshot, it would nearly always show up
as a "difference" from the reference images.

It appears someone might have tried to hide this icon previously, but
the string `backstop-hide-animation` does not appear anywhere else in
the codebase. In any case, we don't need a test-specific class here.