Omit services that are subsets of other services
We have logic in the service picker that gets rid of services with the
same start date, end date, and valid days as another service. However,
there are other cases in which we have two services, one of which is a
subset of another, rather than an exact match. These subsets are
redundant too: omit them.