Allow a single project/project_update teaser in ContentList
Prior to this PR, we were always forcing 2 results when a list of project or project_update teasers were requested -- even when authors explicitly set it to only show 1 result. With this PR, we only enforce the 2-item limit if more than 2 items are requested (as originally spec'd by design).