Fix CMS paragraph order being scrambled
The issue was that `async_with_timeout`, which accepts a list of
functions, was implemented on top of `yield_or_default_many`, which
accepts a map where the values are functions. The map values were being
collected back into a list, but since maps are not ordered, this result
list would not have the same ordering as the input list.

Because the previous approach resulted in this data flow:

list -> map -> list -> Task.yield_many -> list -> map -> list

We can simplify to this:

list -> Task.yield_many -> list