Use correct route colors on line maps
These hard-coded colors were a few years outdated; now that we've added
the `color` field for Route resources in the API, we can use that.

This affects the interactive maps displayed both on the "Info & Maps"
tab on line pages, and when an option is expanded in the Trip Planner.