Fix intermittent react renderer test failure
Previously, in over half of test runs, a random test in `ReactTest`
would fail with a `Task.await` timeout. The cause seems to be that we
were starting a copy of the renderer process within one of the tests,
which somehow interfered with the "real" process. At least, when this
test is deleted, the intermittent failure disappears. The reason may
have something to do with Poolboy, but it's hard to pursue further than
that as Poolboy is written in Erlang and has zero documentation.

Since this test doesn't appear to provide much value, and neither does
the one that stops the process, we can probably safely remove them.
`React.stop` was only used in this test, so it can be deleted as well.

Finally, this change also removes a 10000ms timeout on the `Task.await`
call that could never be reached. This is because the `GenServer.call`
within the task was not given a timeout parameter, and so would time out
after the default of 5000ms. This is also what `Task.await` defaults to,
so removing the parameter aligns the timeout values.