Use more robust live data for finder API tests
The finder API tests as written would occasionally fail due to the
relative paucity of commuter rail schedules and predictions in the
middle of the day. It would be difficult and tedious to mock this out
with our usual DI tricks, and we are hoping to mock out all live data
wholesale soon, so in this case we've replaced one set of live data with
another as a quick fix.