Add test_vcr ExVCR helper
We currently have roughly 3,000 tests, many of which use live data from
the API and/or CMS, which practice we are trying to put an end to. The
ExVCR package allows us to record interactions with remote services for
later use as test fixtures. As such, it's a straightforward way to
eliminate live data use from a test.

However, since we have so many tests, it's important to reduce the
amount of boilerplate involved in converting tests to use ExVCR. This
commit adds a `test_vcr` macro meant to be a dropin replacement for the
normal ExUnit `test` macro, which automatically adds ExVCR use to a test
with no explicit configuration needed.

As one test (`test/site_web/controllers/fare_controller_test.exs`) was
broken at the moment due to a live data change, we also take this
opportunity to use the new helper to fix it.