Fix delayed CR trips showing as "On time"
This addresses an issue where, on stop pages and Transit Near Me,
delayed commuter rail departures would display as "On time" if the
current time was between the scheduled and predicted times. This was
because we requested schedules with a "from now onwards" filter, so the
resulting PredictedSchedule had only a prediction with no schedule, and
the frontend defaults to displaying "On time" in this case. A separate
ticket has been opened regarding this fallback behavior.