Fix blank pages with single-direction routes
The app has not really supported single-direction routes until now, so
this required a significant change to the data modeling: Route direction
names can now be null, indicating there is no service in that direction.

On the backend, this must be determined using RoutePatterns, since there
is no property of the Route itself that indicates which directions are
valid, and RoutePatterns are the only resource that *does* indicate this
and can be included in the same request as a Route. This required some
changes to start including RoutePatterns in Route requests by default.

On the frontend, when we encounter a single-direction route, we always
display data for the sole valid direction and hide UI that would allow
the user to change it. This includes overriding the server-side logic
that would normally pick a default direction based on the time of day.

Interestingly, the existing `Defaults` plug had some code that seemed
intended to support single-direction routes, but it relied on an assign
named `headsigns` that was not assigned or even referenced anywhere else
in the current code.