Refactor line diagram origin/destination logic
We were calculating and passing `isOrigin` and `isDestination` props to
the line diagram stop component, but this data was already available in
the stop itself. A stop is an origin if it `is_beginning?`, and it is a
destination if it `is_terminus?` but is not a beginning.

This also cleans up some fixtures being used to test the line diagram,
which in some cases had `is_beginning?` and `is_terminus?` values that
were inconsistent with their other attributes.