Update Enzyme and bump React patch version
These updates together fix an issue where, when using functional
components, changes to internal component state would not be reflected
in the results of Enzyme functions (e.g. `find`, `exists`, `html`). This
also fixes some tests that unintentionally relied on the bug.

(in the case of SearchBox, the simplest fix was to delete them, as they
were "testing the framework" rather than any logic we wrote)

This also enables the feature of enzyme-to-json where it can act as a
Jest snapshot serializer. This makes it possible to compare a snapshot
just by doing `expect(wrapper).toMatchSnapshot()`, where `wrapper` is an
Enzyme wrapper.