Change "Storm service" copy to "Reduced service"
Currently services with a `typicality` of `unplanned_disruption` are
labelled "Storm service" in the schedule finder, on the assumption that
we would reserve that typicality for winter storms. Given the
possibility of reduced service due to the COVID pandemic, however, here
we change our copy to the more general "Reduced service", which is still
accurate in case of a storm as well.