Selecting variant changes variant query param
Choosing a new variant/route pattern in the menu for same now updates
the variant query param in the address bar, in much the same way that
changing direction does so. This means someone can choose a direction
and variant, and share or bookmark the URL thus generated to immediately
go to the direction and variant selected.