Add hook for tracking/updating where stops are
The stops on the diagram will be drawn next to every stop name, so we need to use the stop name element's offsetTop & offsetHeight via assigning a ref. This is done across many components in the hierarchy so we centralize the logic and create all the refs here.
We also offer a updateAllStops() function that iterates through all the stops to update every location at once.