Add support for special "combo" zones
Currently the 2 combo zones are: "1A-1" and "1A-2". These zones are
treated as Zone 1A for trips between the given station and either North
or South Station, otherwise Zone 1 or Zone 2, respectively.

Use the pattern of the combo zone name to determine all zones, so we can
add additional zones using the same pattern in the future.