feat: authenticate with Docker Hub if credentials are available
In Semaphore, we use DOCKER_USERNAME and DOCKER_PASSWORD for our DockerHub
credentials. This changes the `build.sh` script to log in to DockerHub before
doing other actions if those credentials are available.

The removal of `-x` from the `set` command prevents the password from being
echoed during the build.