fix: ensure that icons have whitespace next to their text
Elixir 1.9 has a small tweak to how templates are parsed:

> [EEx] Consistently trim newlines when you have a single EEx expression per
> line on multiple lines

What this appears to mean for us is that the whitespace between the icon and
the text, when separated by a newline, is combined without whitespace. By
putting on the same link with a space between them, we ensure that the
appropriate spacing is maintained.