feat(ScheduleController.Line.Helpers) Shuttle case
Adds a special case for CR-Fitchburg rail replacement bus service. For the sake of the line diagram, we don’t want to interpret the resulting multiple route patterns on this route as distinct “branches”.

The overarching_branch method added in the prior commit would be able to cover this case, except this particular rail replacement bus service makes an extra stop at Alewife station, which is not covered by the typical CR-Fitchburg route pattern. This commit adds a special case that omits Alewife when computing the overall_branch for this particular line diagram.