Fixing various broken line diagram issues (#806)
* fix(ScheduleController.Line.Helpers): Pick branch

Handle the case of CR-Kingston, which has two route patterns with stops between South Station-Kingston and South Station-Plymouth. These were getting treated as two distinct branches but does not describe a typical "branching" of a line diagram; the latter route pattern merely extends the former by a single stop.

Instead of handling two nearly identical "branches", we select the extended, more complete route pattern for display on the line diagram as a single unbroken, unbranched line.

* feat(ScheduleController.Line.Helpers) Shuttle case

Adds a special case for CR-Fitchburg rail replacement bus service. For the sake of the line diagram, we don’t want to interpret the resulting multiple route patterns on this route as distinct “branches”.

The overarching_branch method added in the prior commit would be able to cover this case, except this particular rail replacement bus service makes an extra stop at Alewife station, which is not covered by the typical CR-Fitchburg route pattern. This commit adds a special case that omits Alewife when computing the overall_branch for this particular line diagram.