fix: Update isCurrentValidService logic
Update to evaluate true for services that are within the current rating, even if the current date does not fall within the service's start/end dates.

This is done to accommodate services that are not active during school vacation days, but would otherwise be active throughout the rating.

If those services were not seen as true here, then on school vacation days we might not have a current service correctly selected in the front end -- this is an edge case where we cannot substitute with the relevant school-vacation-valid service, because those are already filtered out in the back end (because we didn't want to show them in the services dropdown).