feat: Make a wrapper page for the Youth Pass form
The Youth Pass form is the first of a series of reduced fares forms that
will be hosted by SimpliGov. Use a custom wrapper layout to embed an
iframe. The iframe fills the viewport, there are no visible elements
from this project. In the future we will add support for Google
Analytics and potentially additional user analytics tools.

Add configuration support for per-environment reduced fares form URLs.