fix(TripPlan): disambiguate Massport shuttles
Routes, stops, and trips from the Massport GTFS may have identifiers that happen to match those from the MBTA GTFS data. To avoid mixups, when parsing the id we append a "Massport-" prefix. This also means that in certain other places, we can identify Massport shuttles using this same prefix.