feat(TimeHelpers): add displayed_time
The interplay between a scheduled time and a predicted time creates a phenomenon we refer to as a display time. This usually shows the predicted time with the scheduled time as a fallback.

The exact text displayed varies by mode, and also varies by the relationship to "now" (e.g. if the predicted time is right about now, we show "arriving"). This function is lightly refactored from TransitNearMe.format_prediction_time/4, where seconds are removed from the arguments because it can be computed from two of the other arguments.