refactor getTreeDirection to account more termini
This previously assumed all termini were either before or after a merge stop, but in reality termini stops can be on both ends!
This new version counts the number of termini/branch endings and compares the number before and after the merge to determine whether the line diagram is branching inward or outward.