chore: Adjust tests for Summer 2022 rating (#1205)
* chore: Summer 2022 rating replace stop id

stop 71199 will be no more, use a different platform within Park Street station

* chore: Summer 2022 rating tests with ferry service

This is an attempt to make one of our tests more flexible.

Some of our unit tests tend to fail after a rating change because the underlying data returned changes. Often the change is inconsequential, but it makes a test fail anyway because the test itself is usually seeking an overly specific set of IDs, or other specific order of data.

This change attempts to rewrite this test to be flexible enough to assert what we need to, but also pass whether it's the summer rating (which has additional service, and a different headsign) or any other time in the year.