Fix shifting logo on large tablet breakpoint
Specifically on the large tablet breakpoint, the text changing from
`Menu` to `Close` caused the menu button to expand ever so slightly,
causing visual shifting of the MBTA logo. To fix this, I've set a static
width and centered the text, and there appears to be no more visual
shifting when expanding the menu.