Fix line diagram / map crash when spam clicking "Change Direction"
For some lines (noticed on bus routes), the number of stops we get back differs
when reversing directions. Previously, we were looping over these stops and
calling `useRef` for each one. This breaks one of the rules of hooks:
https://reactjs.org/docs/hooks-rules.html

When rapidly clicking the "change directions" button, occasionally the line
diagram and map would blow up with a message about calling hooks out of order,
which is a symptom of the above issue. This change addresses this by instead
using a single ref with `Map` from `stopId`s to `HTMLElement`s, rather creating
a `ref` for each stop.

While I have not been able to reproduce the page crash locally after applying
these changes, I have a feeling this does not address the "true" root cause
(some sort of race condition / stale data? I'm still not 100% sure).

The original reason for digging into this was to address an issue that appears
to only manifest on iOS where clicking the change direction button _once_ would
break the page. Since the effects of both (i.e. the line diagram and map
disappear) are the same, I _hope_ that this will also fix that issue, but I'm
not 100% certain that it will.