Prepare new shared page state
In order for the Schedule Finder(*) to know about the stops for the
currently-selected route variant, we need to share that information between
the Map/Line Diagram and the Schedule Finder.n

This prepares some new page-wide shared-state, which will be shuttled
around via `StateContext`.

(*): technically, ScheduleFinder doesn't actually care about the variant,
it just cares about a route and list of stops (and some other things that
are not important in this context). Instead, we'll be using a new component
in ScheduleLoader to fetch stops based on the currently-selected pattern,
and then passing _those_ into the ScheduleFinder component.