New page structure
Previously, each component on this page was rendered using its own React
"instance", making it complicated to share data across the sections.

This commmit restructures the page so that everything is rendered under one
React instance and with a StateContext.Provider that will be used for
sharing state between the Map/Line Diagram and the Schedule Finder. Each
independent component is mounted into the existing DOM element using React
portals.