refactor: add expanded? prop to accordion component section
Adds expanded? prop to accordion component section, and uses its value to set the initial expanded state.

We also slightly adjust how the interaction and styling work - instead of searching for an absence or presence of the data-accordion-expanded attribute, we explicitly set it to "false" or "true".

Defaults to false / not expanded if not explicitly defined.