refactor: consolidate channels.js & Channel.ts
Combine the completely overlapping functionality into a singular `channels.ts` module. It works mostly the same as before, with minor differences:

* Instead of a new Socket being created multiple times, it's created once and assigned to `window.socket` to be available globally.
* Because we might attempt to join a channel from more than one place, we add a check to avoid joining a channel multiple times.

Adjust `cr-timetable-trains.js` to work with `channels.ts` and update related tests.

Use those functions as the basis for a new generic `useChannel` React hook, which adds/removes event listeners to respond to the custom event created in `channels.ts`.

Adjust the `<Map />` component to use that hook for tracking vehicles, and adjust its tests.