feat(Predictions.StreamParser): Use parent stop ID
Since the predictions stream subscribes to parent stops, use the parent stop when parsing  incoming predictions. Just a nice bit of alignment for our benefit, since we typically use parent stops when working with them on Dotcom.

Also include the platform code that we already have from the API request. We'll need that to display track number later.