Adjust route pattern filtering on backend (again) (#1480)
* feat: add check for empty data in a StopTree

Checking whether `lineState.data.stopTree` is defined is insufficient, because the LineDiagram won't work if a stop tree's nodes and edges are empty lists. This adds a more detailed check with a new helper function, isEmptyTree

* hotfix(Line.Helpers): loosen shape_priority check

Allows usage of route patterns with typicality: 1 that are associated with shape_priority of -1.