cleanup: deprecate old schedule features (#1482)
Our schedule pages used to have a feature where you could pick an origin stop and destination stop. We don't use that anymore but traces of the relevant code were everywhere.

Similar story to the "all_stops" list, which is only now used to list stops in the Timetable (where it's overwritten with a different implementation anyway).

With the removal of those, ExcludedStops is no longer used.