fix: tweak route encoding more
do less in the Algolia modules. do more in the encoder for routes.
it's silly, but Algolia ends up running routes through the encoder twice. this was causing it to fail the second time, because our code didn't expect it. now the new encoded_directions function can handle either case.