Consolidate channel handling (#1481)
* cleanup unused variables

* refactor: consolidate channels.js & Channel.ts

Combine the completely overlapping functionality into a singular `channels.ts` module. It works mostly the same as before, with minor differences:

* Instead of a new Socket being created multiple times, it's created once and assigned to `window.socket` to be available globally.
* Because we might attempt to join a channel from more than one place, we add a check to avoid joining a channel multiple times.

Adjust `cr-timetable-trains.js` to work with `channels.ts` and update related tests.

Use those functions as the basis for a new generic `useChannel` React hook, which adds/removes event listeners to respond to the custom event created in `channels.ts`.

Adjust the `<Map />` component to use that hook for tracking vehicles, and adjust its tests.

* fixup(channels.ts): leave on beforeunload event

Add an extra event listener to leave channels when navigating away from the page. Covers certain transitions that the ‘turbolinks:before-render’ event listener misses.

* fixup: feat: Consolidate channel handling (#1481)

- Add new socketTestHelpers.ts file to the JS coverage ignore list.

* fix(Map): remove check that prevents map updates

- action doesn't have a channel prop to compare

* Revert "fix(Map): remove check that prevents map updates"

This reverts commit e484b528fccce514660ca3c5ad44c4839bf10dcc.

* fixup(Map): remove channel from data type

the reducer will be limited to acting on changes for the given channel by design / using useChannel seems to obviate the need to maintain and compare the channel name.