Subscribe to streaming updates by routeId and stopId (#1476)
* feat: Streaming predictions

Use a DynamicSupervisor for API prediction streams since they need to be
per route—a "firehose" ever everything is not supported by the API.
Start streams when a client subscribes to predictions for a route.
Register streams per route to prevent duplication.

Parse data into pure Prediction structs for the moment. We can think
harder about exactly what data format we want after demonstrating this
proof-of-concept.

* feat: Predictions Phoenix channel

* feat: Parse and save prediction arrival and departure times

Display logic should highly dependent on these, so we should replace the
single generic "time" field with these more accurate distinct values.

* fixup: feat: Parse and save prediction arrival and departure times

* feat: get streaming predictions by route, stop, direction

Adds stop_id and direction_id to the key and to associated functions.

* feat(Predictions.StreamParser): Use parent stop ID

Since the predictions stream subscribes to parent stops, use the parent stop when parsing  incoming predictions. Just a nice bit of alignment for our benefit, since we typically use parent stops when working with them on Dotcom.

Also include the platform code that we already have from the API request. We'll need that to display track number later.

* bugfix(SiteWeb.PredictionsChannel): use "data" event name

It's a bit arbitrary, but this follows the convention set in the Vehicles channel, and allows us to call `channel.on("data", eventHandler)` for both (aka for all channels we have set up so far).

* fixup(PredictionsChannel): emit predictions on join

get whatever predictions are available upon subscribing, remove predictions in the past, push to socket immediately. this way subscribers won't have to wait until the next stream push to get any data.

* fixup(Predictions.StreamParser): handle included

Attributes fetched for related resources (routes, stops, trips) are not reliably available in the context of streaming data. So instead of parsing this from the API response, we take the id and fetch the object via our own `*.Repo` modules.

* fixup: feat(Predictions): Subscribe to streaming updates by routeId and stopId (#1476)

- fix(Predictions.Parser): when determining “first time”, prefer departure time instead of arrival time. This more closely follows V3 API best practice with respect to what time we wish to display to riders.
- fixup(Predictions): update to new child supervisor syntax for `Predictions.PubSub`
- fix(Predictions.PredictionsPubSub): actually add subsequent subscriptions to state
- fix(Predictions.StreamSupervisor): actually use the pid that's registered, rather than the pid the registration is under (!)
- fix(SiteWeb.PredictionsChannel): filter out old predictions in `handle_info` rather than in initialization only
- fix(SiteWeb.PredictionsChannel): remove predictions with no trip information
- fix(SiteWeb.PredictionsChannel): remove predictions with null departure time, as those are likely to be arrivals to terminals that passengers can’t board anyway.
- Make all necessary updates to unit tests

* fixup: remove async: true while using Mock

* fixup(Predictions.PredictionsPubSub): make correction to doc

* fix(Predictions.PredictionsPubSubTest): actually test update

* feat(V3Api.Stream): encode URL

* fixup(SiteWeb.PredictionsChannel): return new predictions immediately

extract helper function to filter the predictions accordingly

* fixup(Predictions.PubSub): correction to comment

---------

Co-authored-by: Matthew Shanley <matthewshanley@littlesecretsrecords.com>