Pick between OTP1 and 2 based on mbta_id cookie
This generates a random value, seeded by the mbta_id cookie, which stays
static per user (until cookies are cleared), to determine which instance
of OTP they should be served. The percentage of traffic being served
OTP2 can be configured with a new env var.

Two new feature flags are also included, which can be used to force a
specific version (OTP2 is picked if both are selected).