Resolve an issue with Massport shuttle labels (#1731)
OTP2 fixed an issue where underscores were removed from GTFS feed IDs,
but dotcom was expecting the Massport ID to always be missing the
underscore that it has in GTFS. This was resulting in Massport shuttles
not being labeled as such and just appearing as regular buses. This adds
another case with the underscore to mark IDs as Massport ones.