fix: adjust form IDs in `form_for`
It turns out Phoenix HTML v3 stopped ignoring the `form_for` id: option when naming form fields. However, we were relying on the as: option to both group the form data and name our form's inputs. Here we just adjust the ID to match the as: option, to maintain the current field naming.

This fixes broken datepicker instantiation on the customer support form and trip planner sidebar.