hotfix(autocomplete): use separate panelContainer
this fixes a Safari bug where the tab order traps keyboard focus in between the autocomplete input and submit buttons. the bug only happened on Safari, when using the React renderer instead of the default Preact renderer.

some CSS refactoring was needed to target elements that are inside the new panelContainer element.