fix(cms): handle case where _format=json is not the last item
The previous documentation indicated that `_format=json` would always be
the final item in the query string. This is not always the case, so our
replacement strategy needs to handle it appearing anywhere.

We do this with a regular expression which can get find the string in
any location, and replace it with either an empty string (if it's at
the beginning or end) or a single `&` if it's in the middle

Test cases are also updated to handle all three positions.